Refactor utils paths and update imports
