+++ backend/backend-ai/utils/extract.js
