+++ frontend/middleware/auth.js
@@ -1,5 +1,5 @@
 const config = require('../config');
-const logger = require('../utils/logger');
+const logger = require('../src/utils/logger');
 
 /**
  * Authentication middleware
