+++ frontend/middleware/rateLimiter.js
@@ -1,6 +1,6 @@
 const rateLimit = require('express-rate-limit');
 const config = require('../config');
-const logger = require('../utils/logger');
+const logger = require('../src/utils/logger');
 
 /**
  * Rate limiting middleware
