+++ frontend/services/openai.service.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const config = require('../config');
-const logger = require('../utils/logger');
+const logger = require('../src/utils/logger');
 
 class OpenAIService {
   constructor() {
